requirements.txt
flask
pandas
matplotlib
openpyxl
requests
streamlit
Pillow
